Data.Matrix.Sparse.Generic

Documentation

class Eq a => Zero a where

Methods

zero :: a

data CSR v a

Compressed Sparse Row (CSR) matrix

Constructors

CSR !Int !Int !(v a) !(Vector Int) !(Vector Int)

type AssocList a = [((Int, Int), a)]

Accessors

length information

dim :: Matrix m v a => m v a -> (Int, Int)

rows :: Matrix m v a => m v a -> Int

Derived methods

Return the number of rows

cols :: Matrix m v a => m v a -> Int

Return the number of columns

Indexing

unsafeIndex :: Matrix m v a => m v a -> (Int, Int) -> a

(!) :: Matrix m v a => m v a -> (Int, Int) -> a

Indexing

takeRow :: Matrix m v a => m v a -> Int -> v a

Extract a row.

takeColumn :: Matrix m v a => m v a -> Int -> v a

Extract a row.

takeDiag :: Matrix m v a => m v a -> v a

Extract the diagonal. Default algorithm is O(min(m,n) * O(unsafeIndex)).

Construction

fromAscAL :: Vector v a => (Int, Int) -> Int -> AssocList a -> CSR v a

Construct CSR from ascending association list. Items must be sorted first by row index, and then by column index.

unsafeFromVector :: Matrix m v a => (Int, Int) -> v a -> m v a

fromVector :: Matrix m v a => (Int, Int) -> v a -> m v a

matrix

Arguments

:: Matrix m v a

=> Int
number of columns

-> [a]
row list

-> m v a

O(m*n) Matrix construction

fromLists :: Matrix m v a => [[a]] -> m v a

O(m*n) Create matrix from list of lists, it doesn't check if the list of list is a valid matrix

fromRows :: Matrix m v a => [v a] -> m v a

O(m*n) Create matrix from rows

empty :: Matrix m v a => m v a

Conversions

flatten :: Matrix m v a => m v a -> v a

Default algorithm is O((m*n) * O(unsafeIndex)).

toRows :: Matrix m v a => m v a -> [v a]

O(m) Return the rows

toColumns :: Matrix m v a => m v a -> [v a]

O(m*n) Return the columns

toList :: Matrix m v a => m v a -> [a]

O(m*n) Create a list by concatenating rows

toLists :: Matrix m v a => m v a -> [[a]]

O(m*n) List of lists