Data.Matrix.Symmetric.Generic

Documentation

data SymMatrix v a

Symmetric square matrix

Constructors

SymMatrix !Int !(v a)

dim :: Matrix m v a => m v a -> (Int, Int)

rows :: Matrix m v a => m v a -> Int

Derived methods

Return the number of rows

cols :: Matrix m v a => m v a -> Int

Return the number of columns

unsafeIndex :: Matrix m v a => m v a -> (Int, Int) -> a

(!) :: Matrix m v a => m v a -> (Int, Int) -> a

Indexing

flatten :: Matrix m v a => m v a -> v a

Default algorithm is O((m*n) * O(unsafeIndex)).

unsafeFromVector :: Matrix m v a => (Int, Int) -> v a -> m v a

fromVector :: Matrix m v a => (Int, Int) -> v a -> m v a

takeRow :: Matrix m v a => m v a -> Int -> v a

Extract a row.

thaw :: (Matrix m v a, PrimMonad s) => m v a -> s (Mutable m (Mutable v) (PrimState s) a)

unsafeThaw :: (Matrix m v a, PrimMonad s) => m v a -> s (Mutable m (Mutable v) (PrimState s) a)

freeze :: (Matrix m v a, PrimMonad s) => Mutable m (Mutable v) (PrimState s) a -> s (m v a)

unsafeFreeze :: (Matrix m v a, PrimMonad s) => Mutable m (Mutable v) (PrimState s) a -> s (m v a)

create :: Matrix m v a => (forall s. ST s (Mutable m (Mutable v) s a)) -> m v a

map :: (Vector v a, Vector v b) => (a -> b) -> SymMatrix v a -> SymMatrix v b

imap :: (Vector v a, Vector v b) => ((Int, Int) -> a -> b) -> SymMatrix v a -> SymMatrix v b

Upper triangular imap, i.e., i <= j

zip :: (Vector v a, Vector v b, Vector v (a, b)) => SymMatrix v a -> SymMatrix v b -> SymMatrix v (a, b)

zipWith :: (Vector v a, Vector v b, Vector v c) => (a -> b -> c) -> SymMatrix v a -> SymMatrix v b -> SymMatrix v c