Data.Matrix.Unboxed

Documentation

type Matrix = Matrix Vector

Accessors

length information

dim :: Context a => Matrix a -> (Int, Int)

rows :: Context a => Matrix a -> Int

cols :: Context a => Matrix a -> Int

Indexing

unsafeIndex :: Context a => Matrix a -> (Int, Int) -> a

(!) :: Context a => Matrix a -> (Int, Int) -> a

takeRow :: Context a => Matrix a -> Int -> Vector a

takeColumn :: Context a => Matrix a -> Int -> Vector a

takeDiag :: Context a => Matrix a -> Vector a

Construction

unsafeFromVector :: Context a => (Int, Int) -> Vector a -> Matrix a

fromVector :: Context a => (Int, Int) -> Vector a -> Matrix a

matrix :: Context a => Int -> [a] -> Matrix a

O(m*n) Matrix construction

fromList :: Context a => (Int, Int) -> [a] -> Matrix a

fromLists :: Context a => [[a]] -> Matrix a

O(m*n) Create matrix from list of lists, it doesn't check if the list of list is a valid matrix

fromRows :: Context a => [Vector a] -> Matrix a

O(m*n) Create matrix from rows

fromColumns :: Context a => [Vector a] -> Matrix a

O(m*n) Create matrix from columns

empty :: Context a => Matrix a

Conversions

flatten :: Context a => Matrix a -> Vector a

toRows :: Context a => Matrix a -> [Vector a]

O(m) Return the rows

toColumns :: Context a => Matrix a -> [Vector a]

toList :: Context a => Matrix a -> [a]

O(m*n) Create a list by concatenating rows

toLists :: Context a => Matrix a -> [[a]]

O(m*n) List of lists

tr :: Context a => Matrix a -> Matrix a

O(m*n) Matrix transpose

subMatrix

Arguments

:: Context a

=> (Int, Int)
upper left corner of the submatrix

-> (Int, Int)
bottom right corner of the submatrix

-> Matrix a

-> Matrix a

O(1) Extract sub matrix

ident :: (Context a, Num a) => Int -> Matrix a

O(m*n) Create an identity matrix

diag

Arguments

:: (Context a, Num a, Foldable t)

=> t a
diagonal

-> Matrix a

O(m*n) Create a square matrix with given diagonal, other entries default to 0

diagRect

Arguments

:: (Context a, Foldable t)

=> a
default value

-> (Int, Int)

-> t a
diagonal

-> Matrix a

O(m*n) Create a rectangular matrix with default values and given diagonal

fromBlocks

Arguments

:: Context a

=> a
default value

-> [[Matrix a]]

-> Matrix a

isSymmetric :: (Context a, Eq a) => Matrix a -> Bool

force :: Context a => Matrix a -> Matrix a

foldl :: Context b => (a -> b -> a) -> a -> Matrix b -> a

Mapping

map :: (Context a, Context b) => (a -> b) -> Matrix a -> Matrix b

imap :: (Context a, Context b) => ((Int, Int) -> a -> b) -> Matrix a -> Matrix b

Monadic mapping

mapM :: (Context a, Context b, Monad m) => (a -> m b) -> Matrix a -> m (Matrix b)

imapM :: (Context a, Context b, Monad m) => ((Int, Int) -> a -> m b) -> Matrix a -> m (Matrix b)

O(m*n) Apply the monadic action to every element and its index, yielding a matrix of results.

mapM_ :: (Context a, Monad m) => (a -> m b) -> Matrix a -> m ()

imapM_ :: (Context a, Monad m) => ((Int, Int) -> a -> m b) -> Matrix a -> m ()

O(m*n) Apply the monadic action to every element and its index, ignoring the results.

forM :: (Context a, Context b, Monad m) => Matrix a -> (a -> m b) -> m (Matrix b)

forM_ :: (Context a, Monad m) => Matrix a -> (a -> m b) -> m ()

Zipping

zipWith :: (Context a, Context b, Context c) => (a -> b -> c) -> Matrix a -> Matrix b -> Matrix c

zipWith3 :: (Context a, Context b, Context c, Context d) => (a -> b -> c -> d) -> Matrix a -> Matrix b -> Matrix c -> Matrix d

zipWith4 :: (Context a, Context b, Context c, Context d, Context e) => (a -> b -> c -> d -> e) -> Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e

zipWith5 :: (Context a, Context b, Context c, Context d, Context e, Context f) => (a -> b -> c -> d -> e -> f) -> Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e -> Matrix f

zipWith6 :: (Context a, Context b, Context c, Context d, Context e, Context f, Context g) => (a -> b -> c -> d -> e -> f -> g) -> Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e -> Matrix f -> Matrix g

izipWith :: (Context a, Context b, Context c) => ((Int, Int) -> a -> b -> c) -> Matrix a -> Matrix b -> Matrix c

izipWith3 :: (Context a, Context b, Context c, Context d) => ((Int, Int) -> a -> b -> c -> d) -> Matrix a -> Matrix b -> Matrix c -> Matrix d

izipWith4 :: (Context a, Context b, Context c, Context d, Context e) => ((Int, Int) -> a -> b -> c -> d -> e) -> Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e

izipWith5 :: (Context a, Context b, Context c, Context d, Context e, Context f) => ((Int, Int) -> a -> b -> c -> d -> e -> f) -> Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e -> Matrix f

izipWith6 :: (Context a, Context b, Context c, Context d, Context e, Context f, Context g) => ((Int, Int) -> a -> b -> c -> d -> e -> f -> g) -> Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e -> Matrix f -> Matrix g

zip :: (Context a, Context b, Context (a, b)) => Matrix a -> Matrix b -> Matrix (a, b)

zip3 :: (Context a, Context b, Context c, Context (a, b, c)) => Matrix a -> Matrix b -> Matrix c -> Matrix (a, b, c)

zip4 :: (Context a, Context b, Context c, Context d, Context (a, b, c, d)) => Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix (a, b, c, d)

zip5 :: (Context a, Context b, Context c, Context d, Context e, Context (a, b, c, d, e)) => Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e -> Matrix (a, b, c, d, e)

zip6 :: (Context a, Context b, Context c, Context d, Context e, Context f, Context (a, b, c, d, e, f)) => Matrix a -> Matrix b -> Matrix c -> Matrix d -> Matrix e -> Matrix f -> Matrix (a, b, c, d, e, f)

Monadic Zipping

zipWithM :: (Context a, Context b, Context c, Monad m) => (a -> b -> m c) -> Matrix a -> Matrix b -> m (Matrix c)

zipWithM_ :: (Context a, Context b, Monad m) => (a -> b -> m c) -> Matrix a -> Matrix b -> m ()

Unzipping

unzip :: (Context a, Context b, Context (a, b)) => Matrix (a, b) -> (Matrix a, Matrix b)

unzip3 :: (Context a, Context b, Context c, Context (a, b, c)) => Matrix (a, b, c) -> (Matrix a, Matrix b, Matrix c)

unzip4 :: (Context a, Context b, Context c, Context d, Context (a, b, c, d)) => Matrix (a, b, c, d) -> (Matrix a, Matrix b, Matrix c, Matrix d)

unzip5 :: (Context a, Context b, Context c, Context d, Context e, Context (a, b, c, d, e)) => Matrix (a, b, c, d, e) -> (Matrix a, Matrix b, Matrix c, Matrix d, Matrix e)

unzip6 :: (Context a, Context b, Context c, Context d, Context e, Context f, Context (a, b, c, d, e, f)) => Matrix (a, b, c, d, e, f) -> (Matrix a, Matrix b, Matrix c, Matrix d, Matrix e, Matrix f)

generate :: Context a => (Int, Int) -> ((Int, Int) -> a) -> Matrix a

Mutable matrix

thaw :: (Context a, PrimMonad s) => Matrix a -> s (MMatrix (PrimState s) a)

unsafeThaw :: (Context a, PrimMonad s) => Matrix a -> s (MMatrix (PrimState s) a)

freeze :: (Context a, PrimMonad s) => MMatrix (PrimState s) a -> s (Matrix a)

unsafeFreeze :: (Context a, PrimMonad s) => MMatrix (PrimState s) a -> s (Matrix a)

create :: Context a => (forall s. ST s (MMatrix s a)) -> Matrix a